refactor: 별점 계산 함수 분리
