--- Frontend/k-rangers-frontend/README.md
@@ -1,5 +1,0 @@
-| Backend | Backend | Frontend | Frontend |
-|---------|---------|----------|----------|
-| <img src="https://avatars.githubusercontent.com/u/132471478?s=100&v=4" width="100px;" alt=""/> | <img src="https://avatars.githubusercontent.com/u/180482872?v=4" width="100px;" alt=""/> | <img src="https://avatars.githubusercontent.com/u/157056310?v=4" width="100px;" alt=""/> | <img src="https://avatars.githubusercontent.com/u/155972130?s=100&v=4" width="100px;" alt=""/> |
-| [**곽희원**](https://github.com/kiw0n) | [**최은서**](https://github.com/eunseo16) | [**이재욱**](https://github.com/Ukja2) | [**현지유**](https://github.com/101nov) |
-| Backend | Backend | Frontend | Frontend |
