+++ Frontend/k-rangers-frontend/src/utils/isOn.jsx
